refactor: nest schema with a type struct and move test definition
