+++ src/schema/mod.rs
@@ -6,6 +6,6 @@ mod definition;
 mod schema;
 mod serde;
 
-pub use schema::definition::Definition;
+pub use schema::definition::{Definition, Type};
 pub use schema::schema::{Schema, SchemaType};
 pub use schema::serde::from_file;
